Processor/pyproject.toml
[tool.poetry]
name = "kcn_processor"
version = "0.1.0"
description = ""
authors = ["''"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.13.0"
aiohttp = { extras = ["speedups"], version = "3.11.9" }
loguru = "0.7.2"
nats-py = "2.9.0"
orjson = "3.10.12"
python-decouple = "3.8"

[tool.ruff]
lint.select = ["ALL"]
lint.ignore = ["FA102", "E501", "PLW0603"]
lint.fixable = ["I", "COM"]
lint.mccabe.max-complexity = 2
lint.pydocstyle.convention = "google"

target-version = "py313"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
